pyproject.toml
[tool.poetry]
name = "tradedataapi"
version = "0.1.0"
description = "A project to ETL HMRC's Trade Data into a PostgreSQL database, and define an API layer on top."
authors = ["Louis Tsiattalou <louis.tsi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
tqdm = "^4.51.0"
wget = "^3.2"
sqlalchemy = "^1.3.20"
pandas = "^1.1.4"
bs4 = "^0.0.1"
requests = "^2.25.0"
psycopg2-binary = "^2.8.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
pytest = "^6.1.2"
jupyter-console = "^6.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
